created an own file for the Slides
